pyproject.toml
[tool.poetry]
name = "real-estate-api"
version = "0.1.0"
description = ""
authors = ["inessochan <iness0chan@gmail.com>"]
readme = "README.md"
packages = [{include = "real_estate_api"}]

[tool.poetry.dependencies]
python = "^3.10"
flask-restful = "^0.3.9"
flask-sqlalchemy = "^3.0.2"
flask-migrate = "^4.0.0"
python-dotenv = "^0.21.0"
argon2-cffi = "^21.3.0"
flask-httpauth = "^4.7.0"
authlib = "^1.1.0"
jwt = "^1.3.1"
flask-limiter = "^2.8.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
